refactor src to packages/slinkity
