+++ packages/slinkity/client/eagerLoader.js
