+++ packages/slinkity/main/defineConfig.js
