+++ packages/slinkity/plugin/applyViteHtmlTransform.js
