+++ packages/slinkity/plugin/index.js
