+++ packages/slinkity/plugin/reactPlugin/1-pluginDefinitions/addPageExtension.js
