+++ packages/slinkity/plugin/reactPlugin/1-pluginDefinitions/addShortcode.test.js
