+++ packages/slinkity/plugin/reactPlugin/1-pluginDefinitions/index.js
