+++ packages/slinkity/plugin/reactPlugin/1-pluginDefinitions/toFormattedDataForProps.js
