+++ packages/slinkity/plugin/reactPlugin/1-pluginDefinitions/toRendererHtml.test.jsx
