+++ packages/slinkity/plugin/reactPlugin/2-pageTransform/toLoaderScript.js
