+++ packages/slinkity/plugin/reactPlugin/index.js
