+++ packages/slinkity/utils/consts.js
