+++ packages/slinkity/utils/fileHelpers.js
