+++ packages/slinkity/utils/resolveConfigFilePath.js
