+++ packages/slinkity/utils/toHtmlAttrString.js
