+++ packages/slinkity/utils/toSlashesTrimmed.js
